LPS-96356 portal-search-ranking-web: Update dependencies
Temporarily adds @babel/plugin-proposal-object-rest-spread dev dependency to fix error in tests:

```
ReferenceError: _objectSpread is not defined
```

Seems to be an issue with v7.5 which is fixed in v7.5.5 so although liferay-npm-scripts is using v7.4.4 it isn't clear why the error is occurring. Adding this depedency for now so we can move forward for the time being.