LPD-58622 Scope locale menu within the component instance
On pages with multiple localized inputs, interacting with
one dropdown could cause it to manipulate the DOM of a different
instance. This would cause UI elements, like the "Manage Translations"
button, to disappear from one dropdown because the update was being
incorrectly applied to another.

The root cause was a global DOM query (`A.one('.palette-container')`)
in the `selectFlag` method. This line was added previously to fix a
bug where the menu reference would become stale when the menu would previously be removed from the DOM.